Fix makefile, f90 matrix input and restructure classes

+++ src/model/types_constants/types.h
@@ -29,7 +29,7 @@ The defined types abstraction for deco stops etc
 
 //Macros
 #define PRES2DEPTH(pressure) (pressure - 1) * 10
-#define DEPTH2PRES(depth) depth / 10 + 1
+#define DEPTH2PRES(depth) static_cast<float> (depth) / 10 + 1
 #define ROUNDSTOP(depth) depth - depth % 3 + 3
 #define MOD(gas, in_deco) static_cast<uint16_t> (DEPTH2PRES(100 * in_deco==true?MAX_DECO_PO2:MAX_BOTTOM_PO2 / gas[0]))
 #define OTU(rate, time, pO2Init) (3*time/11)/(pO2Init+rate*time-pO2Init)*(pow((pO2Init+rate*time-0.5f)/0.5f,11/6)-pow((pO2Init-0.5f)/0.5f,11/6))
